# src/main/java/com/example/onlineLearning/enrollment/model/Enrollment.java
package com.example.onlineLearning.enrollment.model;

import jakarta.persistence.*;


@Entity
@Table(name = "enrollment")
public class Enrollment {
    @Id
    @GeneratedValue(strategy = GenerationType.AUTO)
    private Long id;
    private Long user_id;
    private Long course_id;
    @Enumerated(EnumType.STRING)
    private Status status;

    public Enrollment() {
    }

    public Enrollment(Long user_id, Long course_id, Status status) {
        this.user_id = user_id;
        this.course_id = course_id;
        this.status = status;
    }

    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public Long getUser_id() {
        return user_id;
    }

    public void setUser_id(Long user_id) {
        this.user_id = user_id;
    }

    public Long getCourse_id() {
        return course_id;
    }

    public void setCourse_id(Long course_id) {
        this.course_id = course_id;
    }

    public Status getStatus() {
        return status;
    }

    public void setStatus(Status status) {
        this.status = status;
    }
}