# src/main/java/com/example/onlineLearning/payment/model/Payment.java
package com.example.onlineLearning.payment.model;

import jakarta.persistence.*;

@Entity
@Table(name = "quiz")
public class Payment {
    @Id
    @GeneratedValue(strategy = GenerationType.AUTO)
    private Long id;
    private String amount;

    public Payment() {
    }

    public Payment(Long id, String amount) {
        this.id = id;
        this.amount = amount;
    }

    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public String getAmount() {
        return amount;
    }

    public void setAmount(String amount) {
        this.amount = amount;
    }

}