# src/main/java/com/example/onlineLearning/tracking/model/Tracking.java
package com.example.onlineLearning.tracking.model;

import jakarta.persistence.*;

@Entity
@Table(name = "tracking")
public class Tracking {
    @Id
    @GeneratedValue(strategy = GenerationType.AUTO)
    private Long id;
    private String grade;
    private String average;

    public Tracking() {
    }

    public Tracking(Long id, String grade, String average) {
        this.id = id;
        this.grade = grade;
        this.average = average;
    }

    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public String getGrade() {
        return grade;
    }

    public void setGrade(String grade) {
        this.grade = grade;
    }

    public String getAverage() {
        return average;
    }

    public void setAverage(String average) {
        this.average = average;
    }

}